Update docs, refactor API
