+++ src/Strategies/BodyParameters/GetFromBodyParamTag.php
@@ -1,22 +1,21 @@
 <?php
 
-namespace Mpociot\ApiDoc\Strategies\QueryParameters;
+namespace Mpociot\ApiDoc\Strategies\BodyParameters;
 
 use ReflectionClass;
 use ReflectionMethod;
-use Illuminate\Support\Str;
 use Illuminate\Routing\Route;
 use Mpociot\Reflection\DocBlock;
 use Mpociot\Reflection\DocBlock\Tag;
 use Mpociot\ApiDoc\Strategies\Strategy;
 use Mpociot\ApiDoc\Tools\RouteDocBlocker;
-use Mpociot\ApiDoc\Tools\Traits\ParamHelpers;
 use Dingo\Api\Http\FormRequest as DingoFormRequest;
+use Mpociot\ApiDoc\Tools\Traits\DocBlockParamHelpers;
 use Illuminate\Foundation\Http\FormRequest as LaravelFormRequest;
 
-class GetFromDocBlocks extends Strategy
+class GetFromBodyParamTag extends Strategy
 {
-    use ParamHelpers;
+    use DocBlockParamHelpers;
 
     public function __invoke(Route $route, ReflectionClass $controller, ReflectionMethod $method, array $routeRules, array $context = [])
     {
@@ -36,39 +35,39 @@ class GetFromDocBlocks extends Strategy
                 continue;
             }
 
-            // If there's a FormRequest, we check there for @queryParam tags.
+            // If there's a FormRequest, we check there for @bodyParam tags.
             if (class_exists(LaravelFormRequest::class) && $parameterClass->isSubclassOf(LaravelFormRequest::class)
                 || class_exists(DingoFormRequest::class) && $parameterClass->isSubclassOf(DingoFormRequest::class)) {
                 $formRequestDocBlock = new DocBlock($parameterClass->getDocComment());
-                $queryParametersFromDocBlock = $this->getqueryParametersFromDocBlock($formRequestDocBlock->getTags());
+                $bodyParametersFromDocBlock = $this->getBodyParametersFromDocBlock($formRequestDocBlock->getTags());
 
-                if (count($queryParametersFromDocBlock)) {
-                    return $queryParametersFromDocBlock;
+                if (count($bodyParametersFromDocBlock)) {
+                    return $bodyParametersFromDocBlock;
                 }
             }
         }
 
         $methodDocBlock = RouteDocBlocker::getDocBlocksFromRoute($route)['method'];
 
-        return $this->getqueryParametersFromDocBlock($methodDocBlock->getTags());
+        return $this->getBodyParametersFromDocBlock($methodDocBlock->getTags());
     }
 
-    private function getQueryParametersFromDocBlock($tags)
+    private function getBodyParametersFromDocBlock($tags)
     {
         $parameters = collect($tags)
             ->filter(function ($tag) {
-                return $tag instanceof Tag && $tag->getName() === 'queryParam';
+                return $tag instanceof Tag && $tag->getName() === 'bodyParam';
             })
             ->mapWithKeys(function ($tag) {
-                preg_match('/(.+?)\s+(required\s+)?(.*)/', $tag->getContent(), $content);
+                preg_match('/(.+?)\s+(.+?)\s+(required\s+)?(.*)/', $tag->getContent(), $content);
                 $content = preg_replace('/\s?No-example.?/', '', $content);
                 if (empty($content)) {
-                    // this means only name was supplied
-                    list($name) = preg_split('/\s+/', $tag->getContent());
+                    // this means only name and type were supplied
+                    list($name, $type) = preg_split('/\s+/', $tag->getContent());
                     $required = false;
                     $description = '';
                 } else {
-                    list($_, $name, $required, $description) = $content;
+                    list($_, $name, $type, $required, $description) = $content;
                     $description = trim($description);
                     if ($description == 'required' && empty(trim($required))) {
                         $required = $description;
@@ -77,14 +76,13 @@ class GetFromDocBlocks extends Strategy
                     $required = trim($required) == 'required' ? true : false;
                 }
 
-                list($description, $value) = $this->parseParamDescription($description, 'string');
-                if (is_null($value) && ! $this->shouldExcludeExample($tag)) {
-                    $value = Str::contains($description, ['number', 'count', 'page'])
-                        ? $this->generateDummyValue('integer')
-                        : $this->generateDummyValue('string');
-                }
+                $type = $this->normalizeParameterType($type);
+                list($description, $example) = $this->parseParamDescription($description, $type);
+                $value = is_null($example) && ! $this->shouldExcludeExample($tag)
+                    ? $this->generateDummyValue($type)
+                    : $example;
 
-                return [$name => compact('description', 'required', 'value')];
+                return [$name => compact('type', 'description', 'required', 'value')];
             })->toArray();
 
         return $parameters;
